alexBot\cogs\voiceCommands.py	too-many-branches
Method vcShake of class VoiceCommands  had 20 branches while Pylint recommends having at most 12.
I extracted _get_channel to make the code more structured and solve that.

Method vcShake of class target_autocomplete  had 14 branches while Pylint recommends having at most 12.
I extracted target_autocomplete_get_channel to make the code more structured and solve that.

Please note that the methods _get_channel and target_autocomplete_get_channel are rather close and it might be good to merge them.

Also note that in the prior code of target_autocomplete in case of not finding a channel in
            if channel is None:
                await interaction.response.send_message("No suitable channel to shake into found", ephemeral=True)
                return

None is return
On the other hand in
        if channel is None or interaction.user.voice.channel == channel:
            return [app_commands.Choice(name="err: no suitable shake channel found", value="0")]

There is a more detailed value.
To be more consistent I return the detailed value.